+++ src/chat/actions/addChatItemAction.ts
@@ -4,15 +4,15 @@ import {
   AddMembershipItemAction,
   AddMembershipMilestoneItemAction,
   AddPlaceholderItemAction,
+  AddPollResultAction,
   AddSuperChatItemAction,
   AddSuperStickerItemAction,
   AddViewerEngagementMessageAction,
   LiveChatMode,
-  ModeChangeAction,
-  AddPollResultAction,
   MembershipGiftPurchaseAction,
-  MembershipGiftRedemptionAction,
   MembershipGiftPurchaseTickerContent,
+  MembershipGiftRedemptionAction,
+  ModeChangeAction,
   ModerationMessageAction,
 } from "../../interfaces/actions";
 import {
@@ -38,9 +38,9 @@ import {
   stringify,
   tsToDate,
 } from "../../utils";
-import { parseBadges, parseMembership } from "../badge";
-import { parseAmountText, parseSuperChat } from "../superchat";
-import { parseColorCode, pickThumbUrl } from "../utils";
+import { parseBadges } from "../badge";
+import { parseSuperChat } from "../superchat";
+import { pickThumbUrl } from "../utils";
 
 export function parseAddChatItemAction(payload: YTAddChatItemAction) {
   const { item } = payload;
@@ -126,8 +126,7 @@ export function parseLiveChatTextMessageRenderer(
     renderer.authorPhoto.thumbnails[renderer.authorPhoto.thumbnails.length - 1]
       .url;
 
-  const { isVerified, isOwner, isModerator, membership } =
-    parseBadges(renderer);
+  const badges = parseBadges(renderer);
 
   const contextMenuEndpointParams =
     renderer.contextMenuEndpoint!.liveChatItemContextMenuEndpoint.params;
@@ -151,10 +150,7 @@ export function parseLiveChatTextMessageRenderer(
     authorChannelId,
     authorPhoto,
     message,
-    membership,
-    isVerified,
-    isOwner,
-    isModerator,
+    ...badges,
     contextMenuEndpointParams,
     rawMessage: message, // deprecated
   };
@@ -182,6 +178,7 @@ export function parseLiveChatPaidMessageRenderer(
 
   const message = renderer.message?.runs ?? null;
   const superchat = parseSuperChat(renderer);
+  const badges = parseBadges(renderer);
 
   const parsed: AddSuperChatItemAction = {
     type: "addSuperChatItemAction",
@@ -193,6 +190,7 @@ export function parseLiveChatPaidMessageRenderer(
     authorPhoto,
     message,
     ...superchat,
+    ...badges,
     superchat, // deprecated
     rawMessage: renderer.message?.runs, // deprecated
   };
@@ -201,31 +199,30 @@ export function parseLiveChatPaidMessageRenderer(
 
 // Super Sticker
 export function parseLiveChatPaidStickerRenderer(
-  rdr: YTLiveChatPaidStickerRenderer
+  renderer: YTLiveChatPaidStickerRenderer
 ): AddSuperStickerItemAction {
-  const { timestampUsec, authorExternalChannelId: authorChannelId } = rdr;
+  const { timestampUsec, authorExternalChannelId: authorChannelId } = renderer;
 
   const timestamp = tsToDate(timestampUsec);
 
-  const authorName = stringify(rdr.authorName);
-  const authorPhoto = pickThumbUrl(rdr.authorPhoto);
+  const authorName = stringify(renderer.authorName);
+  const authorPhoto = pickThumbUrl(renderer.authorPhoto);
 
   if (!authorName) {
     debugLog(
       "[action required] empty authorName (super sticker)",
-      JSON.stringify(rdr)
+      JSON.stringify(renderer)
     );
   }
 
-  const stickerUrl = "https:" + pickThumbUrl(rdr.sticker);
-  const stickerText = rdr.sticker.accessibility!.accessibilityData.label;
-  const { amount, currency } = parseAmountText(
-    rdr.purchaseAmountText.simpleText
-  );
+  const stickerUrl = "https:" + pickThumbUrl(renderer.sticker);
+  const stickerText = renderer.sticker.accessibility!.accessibilityData.label;
+  const superchat = parseSuperChat(renderer);
+  const badges = parseBadges(renderer);
 
   const parsed: AddSuperStickerItemAction = {
     type: "addSuperStickerItemAction",
-    id: rdr.id,
+    id: renderer.id,
     timestamp,
     timestampUsec,
     authorName,
@@ -233,14 +230,10 @@ export function parseLiveChatPaidStickerRenderer(
     authorPhoto,
     stickerUrl,
     stickerText,
-    amount,
-    currency,
-    stickerDisplayWidth: rdr.stickerDisplayWidth,
-    stickerDisplayHeight: rdr.stickerDisplayHeight,
-    moneyChipBackgroundColor: parseColorCode(rdr.moneyChipBackgroundColor),
-    moneyChipTextColor: parseColorCode(rdr.moneyChipTextColor),
-    backgroundColor: parseColorCode(rdr.backgroundColor),
-    authorNameTextColor: parseColorCode(rdr.authorNameTextColor),
+    stickerDisplayWidth: renderer.stickerDisplayWidth,
+    stickerDisplayHeight: renderer.stickerDisplayHeight,
+    ...superchat,
+    ...badges,
   };
 
   return parsed;
@@ -259,18 +252,7 @@ export function parseLiveChatMembershipItemRenderer(
   const authorChannelId = renderer.authorExternalChannelId;
   const authorPhoto = pickThumbUrl(renderer.authorPhoto);
 
-  // observed, MODERATOR
-  // observed, undefined renderer.authorBadges
-  const membership = renderer.authorBadges
-    ? parseMembership(renderer.authorBadges[renderer.authorBadges.length - 1])
-    : undefined;
-  if (!membership) {
-    debugLog(
-      `missing membership information while parsing neww membership action: ${JSON.stringify(
-        renderer
-      )}`
-    );
-  }
+  const badges = parseBadges(renderer);
 
   const isMilestoneMessage = "empty" in renderer || "message" in renderer;
 
@@ -297,7 +279,7 @@ export function parseLiveChatMembershipItemRenderer(
       authorName,
       authorChannelId,
       authorPhoto,
-      membership,
+      ...badges,
       level,
       message,
       duration,
@@ -320,7 +302,7 @@ export function parseLiveChatMembershipItemRenderer(
       authorName,
       authorChannelId,
       authorPhoto,
-      membership,
+      ...badges,
       level,
     };
     return parsed;
@@ -475,21 +457,11 @@ export function parseLiveChatSponsorshipsGiftPurchaseAnnouncementRenderer(
   const channelName = header.primaryText.runs[3].text;
   const amount = parseInt(header.primaryText.runs[1].text, 10);
   const image = header.image.thumbnails[0].url;
+  const badges = parseBadges(header);
 
   if (!authorName) {
     debugLog(
       "[action required] empty authorName (gift purchase)",
-      JSON.stringify(renderer)
-    );
-  }
-
-  const membership = parseMembership(
-    header.authorBadges?.[header.authorBadges?.length - 1]
-  )!;
-
-  if (!membership) {
-    debugLog(
-      "[action required] empty membership (gift purchase)",
       JSON.stringify(renderer)
     );
   }
@@ -499,7 +471,7 @@ export function parseLiveChatSponsorshipsGiftPurchaseAnnouncementRenderer(
       id,
       channelName,
       amount,
-      membership,
+      ...badges,
       authorName,
       authorChannelId,
       authorPhoto,
@@ -515,7 +487,7 @@ export function parseLiveChatSponsorshipsGiftPurchaseAnnouncementRenderer(
     timestampUsec,
     channelName,
     amount,
-    membership,
+    ...badges,
     authorName,
     authorChannelId,
     authorPhoto,