+++ src/interfaces/misc.ts
@@ -13,13 +13,19 @@ export const SUPERCHAT_SIGNIFICANCE_MAP = {
 
 export const SUPERCHAT_COLOR_MAP = {
   "4279592384": "blue",
+  "4280191205": "blue",
   "4278237396": "lightblue",
+  "4278248959": "lightblue",
   "4278239141": "green",
+  "4280150454": "green",
   "4294947584": "yellow",
   "4294953512": "yellow",
   "4293284096": "orange",
+  "4294278144": "orange",
   "4290910299": "magenta",
+  "4293467747": "magenta",
   "4291821568": "red",
+  "4293271831": "red",
 } as const;
 /**
  * Components
@@ -31,24 +37,30 @@ export type SuperChatSignificance =
 export type SuperChatColor =
   typeof SUPERCHAT_COLOR_MAP[keyof typeof SUPERCHAT_COLOR_MAP];
 
-export interface SuperChat {
+export type SuperChatColorFields<T extends object> = {
+  [K in keyof T as K extends `${string}Color` ? K : never]: Color;
+};
+
+export type SuperChat<T extends object> = {
   amount: number;
   currency: string;
   color: SuperChatColor;
   significance: SuperChatSignificance;
-  authorNameTextColor: Color;
-  timestampColor: Color;
-  headerBackgroundColor: Color;
-  headerTextColor: Color;
-  bodyBackgroundColor: Color;
-  bodyTextColor: Color;
-}
+} & SuperChatColorFields<T>;
 
 export interface Membership {
   status: string;
   since?: string;
   thumbnail: string;
 }
+
+export interface Badges {
+  isOwner: boolean;
+  isVerified: boolean;
+  isModerator: boolean;
+  membership?: Membership;
+}
+
 /**
  * 0 - 255
  */